refactor user validation into middleware (#422)
