+++ src/app-account.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import errorMiddleware from './util/error-middleware.js';
+import { errorMiddleware } from './util/middlewares.js';
 import validateUser, { validateAuthHeader } from './util/validate-user.js';
 import {
   bootstrap,
@@ -96,5 +96,3 @@ app.get('/validate', (req, res) => {
     res.send({ status: 'ok', data: { validated: true } });
   }
 });
-
-app.use(errorMiddleware);
