+++ src/app-gocardless/app-gocardless.js
@@ -11,7 +11,7 @@ import {
 } from './errors.js';
 import { handleError } from './util/handle-error.js';
 import { sha256String } from '../util/hash.js';
-import validateUser from '../util/validate-user.js';
+import { validateUserMiddleware } from '../util/middlewares.js';
 
 const app = express();
 app.get('/link', function (req, res) {
@@ -20,13 +20,7 @@ app.get('/link', function (req, res) {
 
 export { app as handlers };
 app.use(express.json());
-app.use(async (req, res, next) => {
-  let user = await validateUser(req, res);
-  if (!user) {
-    return;
-  }
-  next();
-});
+app.use(validateUserMiddleware);
 
 app.post('/status', async (req, res) => {
   res.send({
