+++ upcoming-release-notes/422.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [tcrasset]
+---
+
+Refactor user validation into middleware
